+++ amqpstorm/channel.py
@@ -119,11 +119,9 @@ class Channel(BaseChannel):
             message = self._build_message(auto_decode=auto_decode)
             if not message:
                 self.check_for_errors()
-                if break_on_empty:
-                    sleep(IDLE_WAIT * 10)
-                    if not self._inbound:
-                        break
                 sleep(IDLE_WAIT)
+                if break_on_empty and not self._inbound:
+                    break
                 continue
             if to_tuple:
                 yield message.to_tuple()
@@ -163,6 +161,7 @@ class Channel(BaseChannel):
                 reply_text=reply_text),
                 adapter=self._connection
             )
+            self._connection._remove_channel(self.channel_id)
         finally:
             if self._inbound:
                 del self._inbound[:]
@@ -470,3 +469,4 @@ class Channel(BaseChannel):
             except AMQPConnectionError:
                 pass
         self.close()
+        self._connection._remove_channel(self.channel_id)